Autocomplete and add to array_data
When creating a project, it checks if the language passed is in the array_data.json in the S3 bucket, if not it will add it to the array_data.json.

Created autocomplete route that takes type (for now only "languages" will work) and search. It will match the search string against all data in the array_data[languages] and return potential autocompletes.